python/requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.1.0
torch>=2.0.0
tensorboard>=2.14.0
numpy>=1.24.0
pyyaml>=6.0
